lint

+++ tests/flows/native_token_staking.go
@@ -85,7 +85,7 @@ func NativeTokenStakingManager(network interfaces.LocalNetwork) {
 
 		Expect(ver).Should(Equal(uint16(warpMessages.CodecVersion)))
 		Expect(registerValidatorPayload.NodeID).Should(Equal(nodeID))
-		Expect(registerValidatorPayload.Weight).Should(Equal(uint64(weight)))
+		Expect(registerValidatorPayload.Weight).Should(Equal(weight))
 		Expect(registerValidatorPayload.SubnetID).Should(Equal(subnetAInfo.SubnetID))
 		Expect(registerValidatorPayload.BlsPubKey[:]).Should(Equal(blsPublicKey[:]))
 
@@ -131,7 +131,7 @@ func NativeTokenStakingManager(network interfaces.LocalNetwork) {
 		)
 		Expect(err).Should(BeNil())
 		Expect(validatorRemovalEvent.ValidationID[:]).Should(Equal(validationID[:]))
-		Expect(validatorRemovalEvent.StakeAmount.Uint64()).Should(Equal(uint64(weight)))
+		Expect(validatorRemovalEvent.StakeAmount.Uint64()).Should(Equal(weight))
 
 		// Gather subnet-evm Warp signatures for the SetSubnetValidatorWeightMessage & relay to the P-Chain
 		// (Sending to the P-Chain will be skipped for now)
